# openapi.yaml
openapi: 3.0.3
info:
  title: MySelfServe API
  version: 0.1.0
  description: |
    BC Government Income Assistance Self-Service Portal API.

    MySS 2.0 provides self-service capabilities for BC income assistance clients
    to manage service requests, monthly reports, payments, employment plans,
    messages, and account settings. Ministry workers can access client data
    through an admin support view.

    ## Authentication

    Most endpoints require a JWT Bearer token issued via BCeID/IDIR through
    Keycloak OIDC. The token is signed with HS256 and contains claims for
    `sub` (user_id), `role`, `bceid_guid`, and `idir_username`.

    ## Roles

    - **CLIENT** — BCeID-authenticated income assistance clients
    - **WORKER** — IDIR-authenticated ministry staff (includes ADMIN)
    - **ADMIN** — System administrators (inherits WORKER permissions)

    ## Error Handling

    Errors from the upstream Siebel/ICM system are mapped to standard HTTP
    status codes with structured JSON error bodies. A circuit breaker protects
    the application during ICM outages (5 failures → 30s open window → 503).
  contact:
    name: MySS Development Team
  license:
    name: BC Government
    url: https://www2.gov.bc.ca/
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://myss-api-dev.apps.silver.devops.gov.bc.ca
    description: Development (OpenShift)
  - url: https://myss-api-test.apps.silver.devops.gov.bc.ca
...
security:
  - BearerAuth: []
paths:
...
  /employment-plans:
    get:
      operationId: listEmploymentPlans
      tags: [Employment Plans]
      summary: List employment plans
      responses:
        "200":
          description: Employment plans
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EPListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /employment-plans/{ep_id}:
    get:
      operationId: getEmploymentPlan
      tags: [Employment Plans]
      summary: Get employment plan details
      parameters:
        - name: ep_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employment plan details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EPDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /employment-plans/{ep_id}/sign:
    post:
      operationId: signEmploymentPlan
      tags: [Employment Plans]
      summary: Sign an employment plan with PIN
      parameters:
        - name: ep_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EPSignRequest"
      responses:
        "200":
          description: Employment plan signed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EPSignResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Invalid PIN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  # ──────────────────────────────────────────────
  # Account
  # ──────────────────────────────────────────────
...
  /attachments/upload:
    post:
      operationId: uploadAttachment
      tags: [Attachments]
      summary: Upload a file for virus scanning
      description: |
        Upload a file (max 5 MB). Accepted types: PDF, JPEG, PNG, GIF, BMP, TIFF,
        DOC, DOCX, XLS, XLSX. The file is queued for virus scanning; poll the
        status endpoint until the scan completes.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: File uploaded and queued for scanning
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          description: File too large (max 5 MB)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Unsupported file type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attachments/upload/{scan_id}/status:
    get:
      operationId: getScanStatus
      tags: [Attachments]
      summary: Get virus scan status
      parameters:
        - name: scan_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Scan status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanStatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Scan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /internal/av-scan-result:
    post:
      operationId: receiveAvScanResult
      tags: [Attachments]
      summary: Receive antivirus scan result (internal webhook)
      description: Called by the antivirus scanner service. Authenticated via X-Webhook-Secret header.
      security:
        - WebhookSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanResultWebhookRequest"
      responses:
        "200":
          description: Scan result recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: updated
        "401":
          description: Invalid or missing webhook secret
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Scan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Webhook secret not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attachments/sr/{sr_id}/submit:
    post:
      operationId: submitAttachmentToSR
      tags: [Attachments]
      summary: Attach a scanned file to a service request
      parameters:
        - name: sr_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachmentSubmitRequest"
      responses:
        "201":
          description: Attachment linked to service request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachmentSubmitResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Validation error (e.g. scan not clean)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /attachments/sr/{sr_id}/download:
    get:
      operationId: downloadSRAttachment
      tags: [Attachments]
      summary: Download an attachment from a service request
      parameters:
        - name: sr_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File download
          headers:
            Content-Disposition:
              schema:
                type: string
                example: "attachment; filename=\"document.pdf\""
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /attachments/messages/{msg_id}/{attachment_id}/download:
    get:
      operationId: downloadMessageAttachment
      tags: [Attachments]
      summary: Download an attachment from a message
      parameters:
        - name: msg_id
          in: path
          required: true
          schema:
            type: string
        - name: attachment_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File download
          headers:
            Content-Disposition:
              schema:
                type: string
                example: "attachment; filename=\"document.pdf\""
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  # ──────────────────────────────────────────────
  # Admin: Support View
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token signed with HS256. Claims: `sub` (user_id), `role` (CLIENT/WORKER/ADMIN),
        `bceid_guid` (clients), `idir_username` (workers/admins), `exp` (expiry).
    WebhookSecret:
      type: apiKey
      in: header
      name: X-Webhook-Secret
      description: HMAC secret for internal AV scanner webhook
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ServiceUnavailable:
      description: Upstream service (ICM/Siebel) unavailable — circuit breaker open
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Human-readable error message
      required:
        - detail
...
    EPStatus:
      type: string
      enum: [Received, Submitted, PendingSignature]
    ScanStatus:
      type: string
      enum: [PENDING, SCANNING, CLEAN, INFECTED]
...
    EmploymentPlan:
      type: object
      required:
        - ep_id
        - icm_attachment_id
        - status
        - plan_date
        - message_deleted
      properties:
        ep_id:
          type: integer
        message_id:
          type: integer
          nullable: true
        icm_attachment_id:
          type: string
        status:
          $ref: "#/components/schemas/EPStatus"
        plan_date:
          type: string
          format: date
        message_deleted:
          type: boolean
    EPListResponse:
      type: object
      required:
        - plans
      properties:
        plans:
          type: array
          items:
            $ref: "#/components/schemas/EmploymentPlan"
    EPDetailResponse:
      allOf:
        - $ref: "#/components/schemas/EmploymentPlan"
    EPSignRequest:
      type: object
      required:
        - pin
        - message_id
      properties:
        pin:
          type: string
          minLength: 1
        message_id:
          type: integer
    EPSignResponse:
      type: object
      required:
        - ep_id
        - signed_at
      properties:
        ep_id:
          type: integer
        signed_at:
          type: string
          format: date-time

    # ── Account ─────────────────────────────────
...
    UploadResponse:
      type: object
      required:
        - scan_id
        - filename
        - uploaded_at
      properties:
        scan_id:
          type: string
          format: uuid
        filename:
          type: string
        uploaded_at:
          type: string
          format: date-time
    ScanStatusResponse:
      type: object
      required:
        - scan_id
        - status
      properties:
        scan_id:
          type: string
        status:
          $ref: "#/components/schemas/ScanStatus"
        scanned_at:
          type: string
          format: date-time
          nullable: true
    AttachmentSubmitRequest:
      type: object
      required:
        - scan_id
        - filename
      properties:
        scan_id:
          type: string
        filename:
          type: string
    AttachmentSubmitResponse:
      type: object
      required:
        - attachment_id
        - sr_id
        - filename
      properties:
        attachment_id:
          type: string
        sr_id:
          type: string
        filename:
          type: string
    ScanResultWebhookRequest:
      type: object
      required:
        - scan_id
        - status
        - scanned_at
      properties:
        scan_id:
          type: string
        status:
          $ref: "#/components/schemas/ScanStatus"
        scanned_at:
          type: string
          format: date-time

    # ── Admin: Support View ─────────────────────
...